Update migrations and factories to use decimal for all currency columns in the database

+++ database/migrations/2021_09_08_224549_create_orders_table.php
@@ -17,7 +17,7 @@ class CreateOrdersTable extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id');
             $table->string('transaction_id');
-            $table->integer('total');
+            $table->decimal('total', 9, 2);
             $table->timestamps();
         });
     }
